requirements.txt
# cognitive_synergy/requirements.txt
# Python package dependencies for the Cognitive Synergy Model project
# Use pip install -r requirements.txt to install

# Core ML Libraries
torch>=1.10.0,<2.0.0  # PyTorch (adjust version as needed, check CUDA compatibility via pytorch.org)
torchvision>=0.11.1,<1.0.0 # For image transforms (sync with torch version)
# torchaudio>=0.10.0 # Uncomment if needed for audio tasks later

# Model Backbones & Tokenizers
timm>=0.6.12,<0.7.0          # For ViT models (PyTorch Image Models) - Check latest stable version
transformers>=4.20.0,<5.0.0 # For BERT models and tokenizers (Hugging Face) - Check latest stable version

# Configuration & Utilities
PyYAML>=6.0,<7.0           # For loading YAML configuration files
tqdm>=4.60.0,<5.0.0          # For progress bars
numpy>=1.20.0,<2.0.0         # Fundamental package for numerical operations
Pillow>=9.0.0,<10.0.0        # For image loading (PIL fork)

# Experiment Tracking (Optional)
# wandb>=0.12.0,<0.16.0         # Uncomment if using Weights & Biases - Check latest stable version

# Add other specific dependencies as needed for datasets, metrics, etc.
# Example:
# scikit-learn>=1.0 # For certain metrics or data processing
# pandas>=1.3 # For handling data manifests if not using simple JSON
# opencv-python-headless # If needed for specific image processing

